// packages/headless/src/modules/paymentDetails/gateways/types.d.ts
export enum GatewayTypes {
  FREE = -1,
  STORED = 0,
  // ---
  CARD = 1,
  BANK_TRANSFER = 2,
  DIRECT_DEBIT = 3,
  SEPA = 4,
  OFFLINE = 5,
  MOBILE = 6,
  WALLET = 7,
}
// ...
export interface IGateway {
  allow_manual_store: boolean;
  card_types: ICardType[];
  client?: IClient;
  countries?: ICountry[];
  created_at: string;
  currencies: IGatewayCurrency[];
  deleted_at: null | string;
  gateway_provider: IGatewayProvider;
  gateway_provider_id: IGatewayProvider["id"];
  gateway_settings: IGatewaySetting[];
  id: string;
  is_stored: string;
  name: string;
  name_translated: string;
  next_action: { url: string; method: string; fields: { [key: string]: any } };
  oauth_application_access_token_id: number;
  org_id: IOrg["id"];
  payment_instructions: string;
  payment_instructions_translated: string;
  payment_type: IPaymentType;
  provider: string;
  sca_verified: boolean;
  store_on_payment: boolean;
  store_on_payment_force: boolean;
  store_outside_payment: boolean;
  translations: ITranslation[];
  type: GatewayTypes;
  updated_at: string;
  use_frontend_implementation?: boolean;
}

export interface IGatewayProvider {
  auth_type: GatewayAuthType;
  code: string;
  created_at: string;
  display_fields: null;
  external_payment: boolean;
  external_store: boolean;
  fields: IGatewayProviderField[];
  id: string;
  instructions: string;
  logos?: ILogo[];
  name: string;
  name_translated: null | string;
  needs_address: boolean;
  oauth_application_code: string;
  requires_name: boolean;
  short_description: null | string;
  short_description_translated: null | string;
  store_type: GatewayStoreType;
  translations: string[];
  type: GatewayTypes;
  updated_at: string;
}

// --------------------------------------------------------
// Contexts

export interface GatewayContext {
  card?: any;
  elements?: any;
  element?: any;
  renderer?: Function;
  // ---
  basket_id?: string;
  currency?: ICurrency;
  amount?: number;
  gateway?: IGateway;
  renderless?: boolean;
  // ---
  type?: GatewayTypes;
  ctx?: GatewayContext;
  stored_payment_methods?: Array<any>;
  // ---
  renderless?: boolean;
  can_store?: boolean;
  must_store?: boolean;
  must_auto_pay?: boolean;
  // --- Operation
  operation_id?: string;
  // --- UI
  schema?: JsonSchema;
  uischema?: UISchemaElement;
  model?: IBillingDetail;
  // --- Output
  paymentDetails?: any; // will contain the response from Card, as wel las any model data
  // ---
  error?: RequestError;
}